Bump Sidekiq 5.1.3 to 5.2.1
The most significant change in this version is that the default
concurrency has been lowered from 25 to 10 (https://github.com/mperham/sidekiq/issues/3892).
This doesn't affect omnibus-gitlab because the concurrency is controlled via a
setting that defaults to 25 anyway and is passed in via the `-c` command-line
parameter.

However, source installations (including the GDK) will have to either specify
the concurrency in `sidekiq.yml` or use the `-c` option.

Full list of changes: https://github.com/mperham/sidekiq/blob/master/Changes.md