Pushing to a merge request clears the approvals list if the respective project setting is enabled
Previously if the project had approvals_before_merge setting at 0
the approvals would not be cleared even if the "Pushing to a merge
request clears approvals" project setting was enabled.

This commit removes the approvals_before_merge out of the check of
whether we should or not reset the merge request approvals.