Filter any parameters ending with "key" in logs
Rails does a partial match for strings in the filter_parameters
configuration, so the parameter "key" causes "key_id" to be filtered
even though it's a useful parameter for debugging internal API issues.

We now revise this filter to make any parameter ending with "key" is
filtered.

Relates to https://gitlab.com/gitlab-com/gl-infra/production/issues/463