Relax Danger requirements about changed lines
This relaxes the commit rules regarding the number of changed lines.
Previously we required a detailed description if 20 or more lines were
changed. Starting with this commit we change the rules so Danger only
notifies you about writing a detailed description if:

1. A commit changes more than 3 files
2. The same commit changes more than 30 lines

This should make it easier to do a find replace to fix a typo, without
having to repeat what the subject line already says, while still
reminding developers to write a detailed description for new features
and other larger changes.

Fixes https://gitlab.com/gitlab-org/release/framework/issues/13