Pass group IDs to MilestonesFinder
MilestonesFinder expects group ids in group_ids parameter (it does
`Array(params[:group_ids])`), but DashboardGroupMilestone.build_collection
was passing subquery instead. This doesn't work in rails 5 because
this evaluates into array of NULLs in the final query.

Because a separate query is used for loading group IDs both in rails4
and rails5 we can explicitly pass array of IDs to the MilestonesFinder
instead of subquery anyway.

`select.map` is used because `pluck` is failing with:
`undefined method `map' for #<Arel::Nodes::SqlLiteral...>`