Add Geo::ResetChecksumEvent model
These changes introduce a new Geo event that can be triggered to
reset the repositories checksums for a project on secondary nodes.

This event is necessary when you want to recalculate the checksum
on the primary node and make sure that the secondary node will redo
the calculation to avoid stale checksum in the database.