Improve Emoji detection in commit messages
This changes our Danger integration so that it is smarter about
detecting GitLab emoji in commit messages. This is done using a two step
process:

1. We use the old regular expression to determine if a commit message
   _might_ include an emoji.

2. If this regular expression matches, we then check for all existing
   emoji names in the commit message. This might be a wee bit expensive,
   but because we couple it with the previous step it should be rarely
   executed.

This ensures we don't trigger an error when a commit message includes
text such as `:foo:bar:`, which is not a valid Emoji.

Fixes https://gitlab.com/gitlab-org/release/framework/issues/15