CI/review_apps: Make use of external-dns and ephemeral IPs
Don't rely on static IPs for Ingress/LoadBalancers. To do this, we
implement annotations for `external-dns` to be able to populate
endpoints into our `$REVIEW_APPS_DOMAIN`.

This requires the installation of `external-dns` into the cluster
used by CI, which is not a part of this MR.

See #7683 comments for details of installation.