Assert successful sign in after registration
`.has_personal_area?` waits for sign in but does not ensure that it's
successful. Raise an exception if not successful so we can capture
a screenshot at the point of failure

Adds a workaround and code to help troubleshoot why a new user doesn't
stay signed in after loading a project page.