Fix a sanity framework test
On first load, the login page displays the "Change your password" tab,
not the "Sign in" tab. We now assert on the presence of a text that's
always on this page.

Note that in EE, since a license is added in the `#perform_before_hooks`
method, the password is changed so that when we visit the login page,
the "Sign in" tab is displayed as expected. That's why only the CE job
failed to pass in
https://gitlab.com/gitlab-org/gitlab-qa/pipelines/31060920.

Signed-off-by: Rémy Coutable <remy@rymai.me>