Add Geo::CacheInvalidationEvent model
These changes introduce a new Geo event that can be triggered
to invalidate the cache for a specific key on Geo secondaries.

This event is necessary because enabling/disabling features does
not propagate correctly to Geo secondaries. It could be used for
other caches in the future (or existing caches that we currently
set to very short expiration periods on the secondary).