Send a cache invalidation event toggling features
We use the flipper gem to provide feature gates in GitLab backed
with the ActiveSupportCacheStore adapter.

In a Geo setup, the primary and secondary do not share Redis state.
This means that changes to a feature on the primary do not invalidate
the cache on the secondary, leading to inconsistent behaviour.

To fix this send a cache invalidation event via the Geo log cursor
whenever features are changed.