Fixed problem with the content disposition with wiki attachments
Before https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/21362
all the uploads to the wiki where handle by the UploadsController. Now
that they are stored inside the wiki repo we are not using it any more.
Instead, we're using the old code in the WikisController#show.

By default, this code set the content disposition to 'inline',
that's why files like pdf were failing to display.

We're moving from that approach to benefit from the blob logic
and caching them and also mitigate the former error.