Checking out to FETCH_HEAD works better with tags
When the found "branch" is a tag, Git can't checkout to
the tag with the remote name as a prefix. See:

https://gitlab.com/gitlab-org/gitlab-ee/-/jobs/106218414

So instead, since we're always fetching something, we could
just use FETCH_HEAD in this case, and that also works with tags.

Also, we shouldn't run any EE-specific checks for tags:

https://gitlab.com/gitlab-org/gitlab-ee/-/jobs/105906272

It didn't fail but we could clearly see that there were errors,
and there's no reasons to run it for tags.