Pass 'TOP_UPSTREAM_SOURCE_SHA' in trigger-build
This is needed because `GITLAB_VERSION` has a special meaning in
`omnibus-gitlab` triggers: this is the GitLab version to build.

The problem is that `omnibus-gitlab` also has triggers to run QA for an
`omnibus-gitlab` commit, and if we use `GITLAB_VERSION` in that case,
the comment would be posted on the GitLab CE/EE commit (stored in
`GITLAB_VERSION`), whci hwouldn't make any sense.
Thus we need `TOP_UPSTREAM_SOURCE_SHA` to represent the commit on
which we want to leave a comment.

Signed-off-by: Rémy Coutable <remy@rymai.me>