Backport EE fixes to CE checking resolvability
EE branch: 7921-fix-batch-comments-resolution
fixes the checks to know whether discussion will
be resolved upon posting a comment or not. It
covers scenarios where Batch Comments are enabled
and disabled. To minimise risk of conflicts between
CE and EE, the checks have been backported.

this.shouldResolve() will not exist in CE.