Fix documentation for list runners
We are specifying that we will only return the shared runners on two
condition; 1 if project has any available 2 if shared runners are
enabled.

2 is not correct. In the admin UI, when the user press "disable shared
runners" we still show the runners to show what kind of runners are
available when the user enables them. We want the API to mimic this
behavior from the UI.

closes https://gitlab.com/gitlab-org/gitlab-ce/issues/52918