Remove window scroll in features/project/job_spec
We have some tests that execute `execute_script "window.scrollTo(0,0)"`
which is not needed since we are just checking the dom, which end up in
a some slower tests.

Move the `before` block for `when job starts environment` sub contextes
to one block since all of the are the same.

Some formatting fixes to be consistent.

closes https://gitlab.com/gitlab-org/gitlab-ce/issues/52690