Removes experimental label from cluster views
As part of https://gitlab.com/gitlab-org/gitlab-ce/issues/51716, we need
to remove 'experimental' labels from cluster views (show and form), as
well as the promp about cluster configuration, since it won't apply.

These changes were originally introduced on
https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/22011/, but we're
trying to reduce the size of that MR as it's to big to review with
confidence.