Regenerate Gitaly's config.toml each rspec run
This behaviour I removed in 08cdf65b9e1f, but this reverts this specific
line. In case storages, or other options changes which are set by the
Rails environment, this will easier to migrate to.

Storage paths is kept in plural, to support multiple storage in later
release for just testing purposes. But its probable that multi Gitaly
scenarios need to be tested.