Send continue parameter on for cancel_path
In https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/21595
`ContinueParams` was introducted in the `Projects::JobController#cancel`
since `continue` parameter is not present for in `cancel_path` for
`Projects::JobController#show.sjon` the user is being redirect to the
pipeline page, where it should be redirected to the current job page
instead.

Add the `continue` parameter as a query string for `cancel_path`.