Ensure we create the secrets at the right time
In `deploy`, if the previous deployment failed, we delete/cleanup all
the objects related to the release, including secrets. The problem is
that if we create the root password before that, it will be then
recreated during the deploy with a random value!

By creatigng the secret just before actually deplying a new release, we
ensure that it won't be overriden.

Signed-off-by: Rémy Coutable <remy@rymai.me>