Upgrade to Ruby 2.5.3
Attempt to update google-protobuf for migration-paths

Because the one we were using aren't compatible with
Ruby 2.5.3, and it'll be troublesome to switch Ruby.
Upgrading google-protobuf will be much easier.

All of them will need to be updated for Ruby 2.5.3

We remove oj because we don't really need it and it
doesn't compile on 2.5.3 with that version.

Closes https://gitlab.com/gitlab-org/gitlab-ce/issues/41825