Remove display of local Sidekiq process in /admin/sidekiq
In a cloud-native or high-availability configuration, Sidekiq may not
be running on the same node as the Web worker. The output of the `ps`
output does not work in this case and can mislead admins that
Sidekiq isn't running. Remove it in favor of the Sidekiq dashboard,
which shows local processes using the Redis information.

Closes https://gitlab.com/gitlab-org/gitlab-ce/issues/46133