Backport ServiceHook#execute from EE
In EE this method takes an additional argument that specifies the name
of the hook to trigger. There is no particular reason to not backport
this to CE, since by default the behaviour remains the same. By
backporting this code we remove the need for prepending ServiceHook with
a module in EE.