Improve memory performance by reducing dirty pages after fork()
The nakayoshi_fork gem works around the lack of a Ruby compacting
garbage collecting by calling GC.start a few times to "promote" the age
of young objects to the maximum age 3. This reduces the number of dirty
pages and consequently improves copy-on-write behavior with unicorn
and other forking Web application servers.