Add dedicated runner tags to assets job
Ordinarily it would get the `gitlab-org` tag from
`dedicated-no-docs-pull-cache-job`, but adding the `docker` tag
overwrites rather than inherits, so this job has been running on the
underpowered hardware of the shared runners rather than our dedicated
runners.

We're also adding a `high-cpu` tag, which doesn't currently do anything
but will be useful for targeting a runner specifically for this job.