Work around Ruby 2.5.3 crash when a CI job is selected
Ruby 2.5.3 in combination with Rails 5 appears to crash in development
when a CI job is selected for a runner. For some reason, the return
inside the block seems to seg fault. Let's get rid of this since
this isn't a good coding practice in any case.

Relates to https://gitlab.com/gitlab-org/gitlab-ce/issues/54281

Also see https://bugs.ruby-lang.org/issues/15325.