Speed up "show GPG badge" feature specs
Previously, this group of specs would create a new user, add them as a
maintainer, sign them in, load the project's commit list, click a
specific GPG key badge, and then verify the content in the popover.

Now, we make the project public in order to skip the user setup and
login, and load a specific commit page because it loads faster than the
commit list while still showing the GPG badge.