Simplify "changes to verified" GPG feature specs
Previously, we wasted time creating a maintainer and signing them in,
then loaded the entire commit list just to verify the GPG status of a
single commit.

Now, we rely on the project being public so no user needs to be
authenticated, and load just the page for the commit we care about.