Condense the "new_dropdown" layout partial
After converting the test that verifies this partial's behavior from a
feature spec to a view spec, we're able to remove a few CSS classes that
were only used for scoping tests found in the feature spec. All the view
spec needs is the `have_link` matcher which allows us to check for a
specific `href` attribute.