Disable password autocomplete in mirror form fill
Chrome and other browsers ignore autocomplete=false and autocomplete=off
now. This causes the mirroring section to expand whenever a user clicked
"Expand" in the Repository settings.

https://developers.google.com/web/fundamentals/design-and-ux/input/forms/?hl=en
recommends using new-password as the new field to prevent automatic
filling.

It looks like we already do this in app/views/shared/_field.html.haml.

This is a CE backport of
https://gitlab.com/gitlab-org/gitlab-ee/merge_requests/8477.

Closes https://gitlab.com/gitlab-org/gitlab-ce/issues/50812